pyproject.toml
[tool.poetry]
name = "trees-everywhere"
version = "0.1.0"
description = ""
authors = ["Hugo Brilhante <hugobrilhante@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.1"
django-configurations = "~2.4"
django-cache-url = "~3.4.4"
dj-database-url = "~1.2.0"
dj-email-url = "~1.0.6"
gunicorn = "~20.1.0"
psycopg2 = "~2.9.5"
whitenoise = "~6.3.0"
gevent = "~22.10.2"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
model-bakery = "~1.10.1"
coverage = {extras = ["toml"], version = "~7.1.0"}

[tool.black]
line-length = 120

[tool.isort]
multi_line_output = 3
include_trailing_comma = 'True'
force_grid_wrap = 0
use_parentheses = 'True'
ensure_newline_before_comments = 'True'
line_length = 120

[tool.pylint.master]
ignore = [
    "asgi.py",
    "wsgi.py",
    "manage.py"
]

[tool.pylint.messages_control]
disable = [
    "no-member",
    "too-many-ancestors",
    "missing-function-docstring",
    "missing-module-docstring",
    "missing-class-docstring",
    "too-few-public-methods"
]

[tool.coverage.run]
omit =[
  "*migrations/*",
  "*tests/*",
  "*apps.py",
  "*urls.py",
  "*asgi.py",
  "*wsgi.py",
  "*conftest.py",
  "*settings.py",
  "manage.py"
]

[tool.pylint.format]
max-line-length = "120"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
